+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - improved topdata_report table structure
 - Added: command to check for crashed jobs and mark them in the report table
 - Added: option to delete reports with no PID using --delete-no-pid flag
+- Added: UtilThrowable
 
 
 ## [1.1.0] - 2025-02-27