+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3] - 2025-05-04
+### Added
+- IntSet and StringSet helper classes
+
+
 ## [1.1.3] - 2025-03-05
 - improved topdata_report table structure
 - Added: command to check for crashed jobs and mark them in the report table